requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
boto3==1.35.49
botocore==1.35.49
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1725267488082/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725560564262/work
cfgv==3.4.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1728479282467/work
click==8.1.7
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.2
dill==0.3.8
distlib==0.3.9
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1726613473834/work
filetype==1.2.0
fire==0.7.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.67.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1634280454336/work
hpack==4.0.0
huggingface-hub==0.24.7
humanfriendly==10.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
identify==2.6.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1726459485162/work
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.17.0
onnxruntime==1.19.2
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.2.3
patito==0.8.0
pillow==11.0.0
platformdirs==4.3.6
polars==1.12.0
pre-commit==3.8.0
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pybboxes==0.1.6
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1711811537435/work
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1725456139051/work
regex==2024.9.11
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1717057054362/work
requests-toolbelt==1.0.0
roboflow==1.1.48
ruff==0.7.1
s3transfer==0.10.3
safetensors==0.4.5
sahi==0.11.18
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
tokenizers==0.20.1
torch==2.5.0
torchvision==0.20.0
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1722737464726/work
transformers==4.46.0
triton==3.1.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.2
ultralytics==8.3.23
ultralytics-thop==2.0.9
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1726496430923/work
virtualenv==20.27.1
Werkzeug==3.0.6
xxhash==3.5.0
yarl==1.16.0
yolov5==7.0.13
zstandard==0.23.0
